Crate image_sync_operator

image-sync-operator is a Kubernetes operator that synchronizes container images between registries based on the ImageSync custom resource definition (CRD).

Installing the operator is typically done via helm by pulling the chart from the public oci repository. You can also find the charts in the legacy helm registry format here.

To install the operator in single-namespace mode with the default settings you would do the following:

helm repo add apexnw oci://ghcr.io/apexnorthwest/charts
helm install image-sync-operator apexnw/image-sync-operator -n image-sync-operator --create-namespace

To fully customize the configuration you would pass a values.yaml files like so:

---
# Settings for the operator itself. This is the container that runs the operator code and watches for ImageSync CRs.
operator:
  # Image url to pull from
  image: 
    repository: "ghcr.io/apexnorthwest/image-sync-operator"
    tag: "0.1.0"
    pullPolicy: "IfNotPresent"
  # Enable for debug logging. This is not recommended for production use but if you even need to open an issue
  # on the project then we will expect you to provide logs with this enabled.
  debug: false
service_account:
  # Should this chart create a service account?
  # If false, the service account named below must already exist
  create: true
  # Name of the service account to use/create
  # Defaults to release name if not set
  name: ""
# Settings related to the skopeo image used by the operator
# Technically, any image that contains skopeo can be used, but we prefer the official one
# Notably, the ca trust must be in /etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem, so
# if you use an image based on a different OS, it's likely it will have tls problems.
skopeo:
  # Image url to pull from
  image:
    repository: "quay.io/skopeo/stable"
    tag: "latest"
    pullPolicy: "IfNotPresent"
  # Optional CA trust bundle to use when verifying TLS connections to registries
  # This should be a PEM encoded bundle of CA certificates, and will be appended
  # to the system CA trust bundle in the skopeo container. This is generally mandatory when
  # working with private, authenticated registries.
  ca_trust_bundle: |-
    -----BEGIN CERTIFICATE-----
    ***********omitted***********
    -----END CERTIFICATE-----
# Settings related to the operator's access
rbac:
  # Should this chart create the cluster role `image-sync-operator`?
  # This must be set to true if cluster_scoped is true, otherwise it is optional.
  create_cluster_role: false
  # Will this operator need to be cluster scoped?
  # If true, the operator will watch all namespaces and require cluster role permissions.
  cluster_scoped: false
  # When not in cluster scoped mode, which namespaces should the operator watch?
  # If empty, the operator will watch the namespace it is deployed in only.
  # This will be ignored if cluster_scoped is true.
  # The operator service account will need to have a rolebinding in these namespaces to be able to watch for ImageSync CRs.
  # The Role can be copied from the install namespace, or if create_cluster_role is true then you can bind that cluster role into the namespace.
  watched_namespaces:
    - "image-sync-operator"
    - "default"
    - "some-other-application"

You would then create ImageSync CRs in that same namespace (as in this example we’ve installed the operator in restricted mode, as is the default)

The following is an example of an ImageSync:

---
apiVersion: imagesync.apexnw.dev/v1alpha1
kind: ImageSync
metadata:
  name: alpine-latest
  namespace: image-sync-operator
spec:
  source:
    # Full URI of the source image as would be passed to skopeo (minus the docker:// prefix, which is presumed)
    # Note: You cannot use docker short names for images. You must use full docker.io paths, as skopeo cannot know what registry to use for short names. This applies to both source and dest.
    # This can also be an @ hash reference. Destination however, cannot be as you must tag the pushed image to something.
    image: "docker.io/library/alpine:latest"
    # Optional: The name of a Secret in the same namespace of type kubernetes.io/dockerconfigjson used to authenticate to the source registry.
    registryLoginSecret: "dockerhub"
  destination:
    # Full URI of the destination image as would be passed to skopeo (minus the docker:// prefix, which is presumed)
    image: "my-private-registry.example.com/library/alpine:latest"
    # Optional: The name of a Secret in the same namespace of type kubernetes.io/dockerconfigjson used to authenticate to the destination registry.
    registryLoginSecret: "private-registry"
  # Optional: If you want to rerun the sync on a schedule after the initial sync. Uses CronJob format. If unset, the sync only runs once. 
  cronSchedule: "*/15 * * * *"
  # Optional: If true, all architectures will be copied. If false, only the architecture of the operator pod will be copied. Defaults to false. This is the same as passing --all to skopeo.
  allArchitectures: true
  # Optional: If true, the destination image will be tagged with the digest of the source image. If false, the destination image will be tagged with the tag of the source image. Defaults to false. This is the same as passing --preserve-digests to skopeo
  # Notably, if this is set to true, and allArchitectures is set to false, multi-arch images will fail to copy and the sync will fail.
  preserveDigests: true
  # Optional: If you want to pass extra arguments to skopeo, you can do so here. This is a string that will be split on whitespace and passed to skopeo as-is. For example, if you want to pass --insecure-policy to skopeo, you would set this to "--insecure-policy".
  # Note that this is not a list of arguments, but a single string that will be split on whitespace. If it needs to have spaces in an argument (which it shouldn't), you will need to use single quotes inside the string.
  extraSkopeoArguments: ""

This project is not meant to be used as a library or installed via cargo. As a result, all contained functions and modules documented within this site are written for the use of contributors and not users of the operator.

Modules

config 🔒

Config represents the yaml config file consumed by the operator at startup from /config/config.yaml. The main loop uses the read_config_file() function to read the config file and return a Config struct.

cronjobs 🔒

This module contains the helper functions to manage the CronJobs used by scheduled image syncs.

imagesync 🔒

This module contains the ImageSync object type and all configuration required to make serde behave.

jobs 🔒

Job management for ImageSync operator. This only controls the one-shot and initial jobs.

reconciler 🔒

Enums

Error

Catchall error type for the operator.

Statics

CONFIG

Global config object. This is ready lazily from /config/config.yaml and is used to set global config options in the operator.

Functions

error_policy 🔒

This function is called when the reconciler encounters an error. In this case, we simply retry the reconcile after 30 seconds. Because the reconciler is idempotent, this is safe to do. The error is logged by the controller runtime.

main 🔒

Entrypoint for the operator. This main function sets up the leader election, watches for ImageSync CRD changes, and calls the reconciler when a change is detected. It uses tokio as the async runtime and is not meant to be called directly.

reconcile 🔒

This reconcile funtion is the main logical driver of the operator. It is called whenever an event is generated related to an ImageSync CRD. This runs as a single listener in global or single-namespace mode. In multi-namespace mode it will run as a listener on each namespace it monitors. Notably, this function usually only performs a single action when called, even if multiple actions are called for. This is because we want the function to be as idempotent as possible. It also ensures that no matter what state the ImageSync is in, it will always progress towards the correct state.

Type Aliases

Result