image_sync_operator::cronjobs

Function get_cronjob_for_imagesync

pub async fn get_cronjob_for_imagesync(
    obj: &ImageSync,
    client: &Client,
) -> Result<Option<CronJob>, Box<dyn Error>>

Get the CronJob associated with the given ImageSync CR, if it exists.

Returns a CronJob object if it finds one, None if it does not, or Errors out if there are more than one job. This should never happen unless something has gone quite wrong.