image_sync_operator

Module reconciler

Functions

acceptance_checks

Run acceptance checks on the ImageSync spec.

has_config_changed

Checks if the spec of the ImageSync CR has changed since the last time it was applied. Returns true if it has changed, false otherwise.

reset_to_not_accepted

When the ImageSync spec has changed, we need to reset the status to not accepted.

update_status

Take a list of new status values and update the CR. This is a common task in the reconciler loop so we abstract it out.